+++ tools/zndq/CHANGELOG.md
@@ -1,7 +1,8 @@
-# *JREq* changelog
+# *ZNDq* changelog
 
 2018-01-17:
 - `init(...)` receives now the location where the locale and configuration will be found as parameter,
+- handling array of strings,
 
 2017-12-20:
 - the wrapper can now be loaded more than once; two *addon*s based on this wrapper can now be used together,